Combobox finalizado previsao compras

+++ previsao_compras/previsao_compras.py
@@ -40,7 +40,6 @@ def setarData(tipo_requisicao):
             produtosComposicao = connection.getProdutosComposicao(dtInicioFormatada, dtFimFormatada)
             composicaoSemiAcabados = connection.getCompSemiAcabados(dtInicioFormatada, dtFimFormatada)
             ajustes = connection.getAjustes(dtInicioFormatada, dtFimFormatada)
-            print(dtInicioFormatada, dtFimFormatada)
         else:
             n_dt_inicio = int(dtInicioFormatada) - 10000
             n_dt_fim = int(dtFimFormatada) - 10000
@@ -67,11 +66,6 @@ def setarData(tipo_requisicao):
     else:
         tabelas.criarTabela(secondFrame)
         return None
-
-def checarEventosNaLista():
-    for evento in tabelas.tabelaSemana.get_children():
-        print(tabelas.tabelaSemana.item(evento))
-
 
 def formatarDataPedido(data):
     milliseconds_since_epoch = data
@@ -159,28 +153,53 @@ def criarDictSemiAcabados(acabados, semiAcabados, estoque):
     listaFormatada = formatarListaSemiAcabados(listaJson, estoque)
     return listaFormatada
 
+######## SELECIONAR CLASSIFICAÇÃO PRODUTO RAPHAEL 08/05/2024 ########
+
+def filtrarListas(tipoFiltro, listaCompleta):
+    if listaCompleta == None:
+        return None
+    elif listaCompleta == 0:
+        return 0
+    else:
+        listaFiltrada = list(filter(lambda produto:produto['IDX_CLASSIFICACAO'] == tipoFiltro 
+                                    or tipoFiltro in produto['IDX_CLASSIFICACAO'], listaCompleta))
+        return listaFiltrada
+
+######## ######## ######## ################ ################ ########
+
 def inserirNaLista():
     produtos = setarData('ano-atual')
+    valorSelecionado = combo.get()
+    
     
     if produtos == None:
         messagebox.showinfo('Data inválida', 'Periodo selecionado inválido')
     elif produtos == 0:
         messagebox.showinfo('Lista vazia', 'Não há eventos nesse período de tempo')    
     else:
-        produtosOrdenados = sorted(produtos, key=lambda p:p['nomeProdutoComposicao'], reverse=True)
-        
-        tabelas.table.delete(*tabelas.table.get_children())
-        for p in produtosOrdenados:
-            id = p['idProdutoComposicao']
-            nome = p['nomeProdutoComposicao']
-            classificacao = p['classificacao']
-            estoque = p['estoque']
-            unidadeEstoque = p['unidadeEstoque']
-            totalProducao = p['totalProducao']
-            unidade = p['unidade']
-            data = (id, nome, classificacao, estoque, unidadeEstoque, totalProducao, unidade)
-            tabelas.table.insert(parent='', index=0, values=data)
-
+        if valorSelecionado == 'Todos os produtos':
+            return produtos
+        else:
+            produtosFiltrados = filtrarListas(valorSelecionado, produtos)
+            
+            if produtosFiltrados != None or produtosFiltrados != 0:
+                produtosFiltrados = sorted(produtosFiltrados, key=lambda p:p['nomeProdutoComposicao'], reverse=True)
+                tabelas.table.delete(*tabelas.table.get_children())
+                
+                
+                for p in produtosFiltrados:
+                    id = p['idProdutoComposicao']
+                    nome = p['nomeProdutoComposicao']
+                    linha = p['classificacao']
+                    classificacao = p['IDX_CLASSIFICACAO']
+                    estoque = p['estoque']
+                    unidadeEstoque = p['unidadeEstoque']
+                    totalProducao = p['totalProducao']
+                    unidade = p['unidade']
+                    data = (id, nome, linha, classificacao, estoque, unidadeEstoque, totalProducao, unidade)
+                    tabelas.table.insert(parent='', index=0, values=data)
+            else:
+                return 'Nenhum produto neste periodo.'
 
 def gerarPlanilha():
     produtos = setarData('ano-atual')
@@ -249,58 +268,30 @@ dtFim.grid(row=2, column=1, padx=(50, 0), pady=5, sticky="w")
 btn_obter_data = Button(secondFrame, text="Mostrar lista", bg='#C0C0C0', font=("Arial", 16), command=inserirNaLista)
 btn_obter_data.grid(row=5, column=0, columnspan=2, padx=(80, 0), pady=2, sticky='nsew')
 
-#row 7 --> Tabela composição acabados
 
 btn_abrir_janela = Button(secondFrame, text="Ver qtd. ano anterior", bg='#C0C0C0', font=("Arial", 16), command=lambda: verQtdAnoPassado(tabelas.table))
 btn_abrir_janela.grid(row=8, column=0)
 
-# btn_mostrar_eventos = Button(secondFrame, text="Ver todos os eventos", bg='#C0C0C0', font=("Arial", 16), command= lambda:verTodosEventos(ajustes_periodo, tabelas.table))
-# btn_mostrar_eventos.grid(row=8)
-
-#row 10 --> Tabela composição semi-acabados
 
 btn_obter_data = Button(secondFrame, text="Gerar Planilhas Excel", bg='#C0C0C0', font=("Arial", 16), command=gerarPlanilha)
 btn_obter_data.grid(row=17, column=0, columnspan=2, padx=(80, 0), pady=(10, 30), sticky='nsew')
 
 
-####################################################
-#PÁGINA 2
-####################################################
-
-# page2 = Frame(notebook)
-# notebook.add(page2,text='Página 2')
-
-# lb1 = Label(page2, text='I am page 2')
-# lb1.grid(pady=20)
-
-# hora_ultima_checagem = Label(page2, text='', bg='#C0C0C0', font=("Arial", 16))
-# hora_ultima_checagem.grid(row=0, column=0)
-
-# mensagem_banco = Label(page2, text='', font=("Arial", 16))
-# mensagem_banco.grid(row=1, column=0)
-
-# dt_inicio_semana = Label(page2, text="De:", font=("Arial", 14))
-# dt_inicio_semana.grid(row=2, padx=(0, 190), column=0, sticky="e")
-
-# dt_inicio_semana = DateEntry(page2, font=('Arial', 12), width=22, height=20, background='darkblue', foreground='white', borderwidth=2, date_pattern='dd/mm/yyyy')
-# dt_inicio_semana.grid(row=3, column=0, padx=(150, 0), pady=5, sticky="e")
-
-# dt_fim_semana = Label(page2, text="Até:", font=("Arial", 14))
-# dt_fim_semana.grid(row=2, column=1, padx=(50, 0), pady=5, sticky="w")
-
-# dt_fim_semana = DateEntry(page2, font=('Arial', 12), width=22, height=20, background='darkblue', foreground='white', borderwidth=2, date_pattern='dd/mm/yyyy')
-# dt_fim_semana.grid(row=3, column=1, padx=(50, 0), pady=5, sticky="w")
-
-# btn_pedidos_semana = Button(page2, text="Ver pedidos meio semana", bg='#C0C0C0', font=("Arial", 16), command= lambda: inserirTabelaTeste('btn'))
-# btn_pedidos_semana.grid(row=4)
-# #verTodosEventos
-# btn_mostrar_eventos = Button(page2, text="Ver todos os eventos", bg='#C0C0C0', font=("Arial", 16), command= lambda:verTodosEventos(ajustes_meio_semana, tabelas.tabelaSemana))
-# btn_mostrar_eventos.grid(row=7)
-
-#tabelas.criarTabelaMeioSemana(page2)
+#Combo box classificação
+opcoes = ['Todos os produtos','Acessórios', 'Acompanhamento', 'Bebidas', 'Bolo', 'Buffet', 'Cafeteria', 'Carnes', 
+          'Compras Diversas', 'Congelados', 'Conservas', 'Copos', 'Descartáveis', 'Diversos', 'Doces', 
+          'Doces Terceirizados', 'embalagem', 'Equip.Segur/EPI/Unif', 'Frete', 'Frios', 'Frutas', 
+          'Higiene / Limpeza', 'Hortifrutigranjeiros', 'Laticínios', 'Louças', 'Manutenção', 'Massa', 
+          'Materiais', 'Material Escritorio', 'Mercearia', 'Mesa de Antepastos', 'Mesa de Guloseimas', 
+          'Pães e Bolos', 'Patê', 'Peças Inox', 'Peixe', 'Pescados', 'Petiscos', 'Petiscos Frios', 'Produtos revenda', 
+          'Recheios e Massas', 'Refrigerados', 'Salada', 'Sobremesa', 'Sobremesa Terceiriza', 
+          'Sorveteria', 'Utensilios']
+opcaoSelecionada = StringVar()
+opcaoSelecionada.set('Todos os produtos')
+combo = ttk.Combobox(secondFrame, values=opcoes, textvariable=opcaoSelecionada)
+combo.grid(row=4, padx=(160, 100), columnspan=2, sticky='nsew')
+combo.bind("<<ComboboxSelected>>")
 tabelas.criarTabela(secondFrame)
-
-#consultarAttBanco()
 
 root.mainloop()
